# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Educ-Map API Documentation",
        "description": "API for higher education institutions directory in Madagascar",
        "contact": {
            "name": "Educ-Map Support",
            "email": "contact@educ-map.mg"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "https://educ-map.lndo.site",
            "description": "Educ-Map API Server"
        },
        {
            "url": ""
        }
    ],
    "paths": {
...
        "/api/grades": {
            "get": {
                "tags": [
                    "Grades"
                ],
                "summary": "Get a list of all grades",
                "description": "Display a listing of the grades.",
                "operationId": "937082441649d3c80b32572bdcd1908b",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search by grade name (partial match)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "description": "Field to sort by (name, level, created_at)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "level",
                            "enum": [
                                "name",
                                "level",
                                "created_at"
                            ]
                        }
                    },
                    {
                        "name": "sort_direction",
                        "in": "query",
                        "description": "Sort direction (asc or desc)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "asc",
                            "enum": [
                                "asc",
                                "desc"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of grades",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/GradeResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Grades"
                ],
                "summary": "Create a new grade",
                "description": "Store a newly created grade in storage.",
                "operationId": "09ba8a5b3973d399754f95ab9d83b9ab",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "level": {
                                        "type": "integer",
                                        "maximum": 10,
                                        "minimum": 1,
                                        "nullable": true
                                    },
                                    "description": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Grade created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/GradeResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input data"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/grades/{id}": {
            "get": {
                "tags": [
                    "Grades"
                ],
                "summary": "Get a specific grade",
                "description": "Display the specified grade.",
                "operationId": "336255077855ed76091fbce88d6d11df",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the grade to retrieve",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Grade details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/GradeResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Grade not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Grades"
                ],
                "summary": "Update a grade",
                "description": "Update the specified grade in storage.",
                "operationId": "945be0e2718fba53ecb91c12eca274c1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the grade to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "maxLength": 255
                                    },
                                    "level": {
                                        "type": "integer",
                                        "maximum": 10,
                                        "minimum": 1,
                                        "nullable": true
                                    },
                                    "description": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Grade updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/GradeResource"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input data"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Grade not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Grades"
                ],
                "summary": "Delete a grade",
                "description": "Remove the specified grade from storage.",
                "operationId": "f3957d4c8e9fedc3599185c50b5fdc24",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the grade to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Grade deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Grade deleted successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Grade not found"
                    },
                    "409": {
                        "description": "Conflict - Grade has associated program offerings"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "GradeResource": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Licence"
                    },
                    "level": {
                        "type": "integer",
                        "example": 3
                    },
                    "description": {
                        "type": "string",
                        "example": "Bachelor's degree level",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "Enter token in format (Bearer <token>)"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}